critical package count is shown
