+++ lunatrace/bsl/common/src/group-vulnerabilities-by-package.ts
@@ -100,19 +100,3 @@ function mergeExistingFindingIntoPackage<F extends Finding>(vulnPackage: Vulnera
     });
   }
 }
-
-interface FindingForCounting {
-  severity: string;
-  purl: string;
-}
-
-export function countCriticalVulnerabilities(findings: FindingForCounting[]): number {
-  const criticalPackagePurls = new Set<string>();
-  findings.forEach((finding) => {
-    // TODO (cthompson) getCvssVectorFromSeverities needs to be moved into the common folder
-    if (finding.severity === 'Critical' && finding.purl) {
-      criticalPackagePurls.add(finding.purl);
-    }
-  });
-  return criticalPackagePurls.size;
-}
