+++ lunatrace/bsl/frontend/src/api/graphql/getProject.graphql
@@ -45,6 +45,12 @@ query GetProject($project_id: uuid!, $build_limit: Int = 10) {
                 severity
                 locations
                 vulnerability {
+                    severities {
+                        id
+                        source
+                        type
+                        score
+                    }
                     ignored_vulnerabilities(where: {project_id: {_eq: $project_id}}) {
                         locations
                     }
@@ -116,6 +122,12 @@ query GetProject($project_id: uuid!, $build_limit: Int = 10) {
                     summary
                     source
                     source_id
+                    severities {
+                        id
+                        source
+                        type
+                        score
+                    }
                     guide_vulnerabilities {
                         guide {
                             id
